app/requirements.txt
alembic==1.11.1
fastapi==0.95.2
fastapi-sqlalchemy==0.2.1
psycopg2==2.9.6
pydantic==1.10.8
uvicorn==0.22.0
python-dotenv==1.0.0
geopy==2.3.0
fastapi_filter==0.6.0
celery==5.2.7
redis==4.5.5

pytest==7.3.1
httpx==0.24.1